chore: register message server in module.go
