+++ x/minfee/types/events.go
@@ -7,11 +7,3 @@ func NewUpdateMinfeeParamsEvent(authority string, params Params) *EventUpdateMin
 		Params: params,
 	}
 }
-
-// NewEventUpdateMinfeeParams creates a new instance of EventUpdateMinfeeParams
-func NewEventUpdateMinfeeParams(signer string, params Params) *EventUpdateMinfeeParams {
-	return &EventUpdateMinfeeParams{
-		Signer: signer,
-		Params: params,
-	}
-}
